third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/iframe-popup.https_3-4-expected.txt

This is a testharness.js-based test.
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same origin popup with window_open
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same origin popup with anchor
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same origin popup with form
FAIL COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same origin iframe, same site popup with window_open assert_equals: Popup name is cleared? expected "true" but got "false"
FAIL COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same origin iframe, same site popup with anchor assert_equals: Popup name is cleared? expected "true" but got "false"
FAIL COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same origin iframe, same site popup with form assert_equals: Popup name is cleared? expected "true" but got "false"
Harness: the test ran to completion.
